move pcompass/Workbench/Module.ts to pcompass
